requirements.dev.txt
requests>=2
uvicorn>=0.18.2
pep8-naming==0.13.1
flake8==4.0.1
flake8-annotations==2.9.1
flake8-annotations-coverage==0.0.6
flake8-bugbear==22.7.1
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.10.0
flake8-docstrings==1.6.0
flake8-encodings==0.5.0.post1
flake8-expression-complexity==0.0.11
flake8-fastapi==0.7.0
flake8-helper==0.2.1
flake8-import-order==0.18.1
flake8-multiline-containers==0.0.19
flake8-newspaper-style==0.3.0
flake8-pie==0.15.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-return==1.1.3
flake8-scream==0.1.0
flake8-simplify==0.19.3
flake8-unused-arguments==0.0.11
flake8-useless-assert==0.4.3
flake8-variables-names==0.0.5